refactor: MainPage path changed
