+++ frontend/src/app/layouts/MainLayout/MainLayout.tsx
@@ -5,7 +5,7 @@ import { Outlet } from 'react-router'
 import { MainHeader } from '@base/widgets'
 
 import { LayoutContainer } from './MainLayout.styles'
-import { MainPage } from '@base/pages/MainPage/ui/MainPage'
+import { MainPage } from '@base/pages'
 
 export const MainLayout = () => (
   <LayoutContainer>
